feat: update routes objects

+++ backend/app/Http/Controllers/API/AugumentedObjectController.php
@@ -5,11 +5,13 @@ namespace App\Http\Controllers\API;
 use Illuminate\Http\Request;
 use App\Http\Controllers\API\BaseController as BaseController;
 use App\Models\AugumentedObject;
-use Validator;
+use Illuminate\Validation\Validator;
 use App\Http\Resources\AugumentedObjectResource;
+use Illuminate\Support\Facades\File;
      
 class AugumentedObjectController extends BaseController
 {
+
     /**
      * Display a listing of the resource.
      *
@@ -29,7 +31,7 @@ class AugumentedObjectController extends BaseController
      */
     public function store(Request $request)
     {
-        $input = $request->all();
+        $input = $request->except('ar_file');
      
         $validator = Validator::make($input, [
             'name' => 'required',
@@ -41,6 +43,13 @@ class AugumentedObjectController extends BaseController
         }
      
         $augumentedObject = AugumentedObject::create($input);
+        
+        $augumentedObject->filehash=hash_file('md5', $request->ar_file);
+        $augumentedObject->size = $request->ar_file->getSize();
+        $augumentedObject->extension = $request->ar_file->getClientOriginalExtension();
+        $augumentedObject->path = $request->ar_file->storeAs('objetos/' . $augumentedObject->id, $augumentedObject->filehash.'.'. $augumentedObject->extension);
+        $augumentedObject->save();
+
      
         return $this->sendResponse(new AugumentedObjectResource($augumentedObject), 'AugumentedObject created successfully.');
     } 
@@ -71,11 +80,11 @@ class AugumentedObjectController extends BaseController
      */
     public function update(Request $request, AugumentedObject $augumentedObject)
     {
-        $input = $request->all();
+        $input = $request->except('ar_file');
      
         $validator = Validator::make($input, [
             'name' => 'required',
-            'initial' => 'required'
+            'description' => 'required'
         ]);
      
         if($validator->fails()){
@@ -83,8 +92,17 @@ class AugumentedObjectController extends BaseController
         }
      
         $augumentedObject->name = $input['name'];
-        $augumentedObject->initial = $input['initial'];
-        $augumentedObject->save();
+        $augumentedObject->description = $input['description'];
+
+        if (isset($request->ar_file) && $request->ar_file->getSize() > 0) {
+            $augumentedObject->filehash=hash_file('md5', $request->ar_file);
+            $augumentedObject->size = $request->ar_file->getSize();
+            $augumentedObject->extension = $request->ar_file->getClientOriginalExtension();
+            $augumentedObject->path = $request->ar_file->storeAs('objetos/' . $augumentedObject->id, $augumentedObject->filehash.'.'. $augumentedObject->extension);
+            
+        }
+
+        $augumentedObject->update();
      
         return $this->sendResponse(new AugumentedObjectResource($augumentedObject), 'AugumentedObject updated successfully.');
     }
@@ -97,8 +115,21 @@ class AugumentedObjectController extends BaseController
      */
     public function destroy(AugumentedObject $augumentedObject)
     {
-        $augumentedObject->delete();
+
+        if (File::deleteDirectory(storage_path('app/files/objeto/') . $augumentedObject->id)){
+            $augumentedObject->delete();
+            return $this->sendResponse([], 'AugumentedObject deleted successfully.');
+
+        } else 
+            return $this->sendError('Deleted Error', 'Could not delete file', 400);
      
-        return $this->sendResponse([], 'AugumentedObject deleted successfully.');
+    }
+
+    public function download(AugumentedObject $augumentedObject)
+    {
+        // Check if file exists in app/storage/file folder
+
+        $file_path = storage_path('app/files/') . $augumentedObject->url_download;
+        return response()->download($file_path);
     }
 }
